pyproject.toml
[project]
name = "euspeedrunmaprenders"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts==0.2.2",
    "av==16.1.0",
    "beautifulsoup4==4.14.3",
    "certifi==2026.2.25",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "cloup==3.0.8",
    "decorator==5.2.1",
    "glcontext==3.0.0",
    "idna==3.11",
    "isosurfaces==0.1.2",
    "manim==0.20.1",
    "manimpango==0.6.1",
    "mapbox-earcut==2.0.0",
    "markdown-it-py==4.0.0",
    "mdurl==0.1.2",
    "moderngl==5.12.0",
    "moderngl-window==3.1.1",
    "networkx==3.6.1",
    "numpy==2.4.2",
    "pillow==12.1.1",
    "pycairo==1.29.0",
    "pydub==0.25.1",
    "pyglet==2.1.13",
    "pyglm==2.8.3",
    "pygments==2.19.2",
    "pyyaml>=6.0.3",
    "requests==2.32.5",
    "rich==14.3.3",
    "scipy==1.17.1",
    "screeninfo==0.8.1",
    "skia-pathops==0.9.2",
    "soupsieve==2.8.3",
    "srt==3.5.3",
    "svgelements==1.9.6",
    "tqdm==4.67.3",
    "typing-extensions==4.15.0",
    "urllib3==2.6.3",
    "watchdog==6.0.0",
]
